python-services/requirements.txt
langchain-docling
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
requests
aiofiles
